Set correct userbrick references in actions (#1792)
Fix copying/backpacking of userbricks
Unpack sprite and userbrick variables only if variable with this name is not already existing
Fix loading of Physics bricks from backpack.json
Fix ShowVariable & HideVariable Brick display of variables with same name
Refactoring
Fix correcting of IfLogicBrick-References

+++ catroid/src/org/catrobat/catroid/io/StorageHandler.java
@@ -621,16 +621,25 @@ public final class StorageHandler {
 
 		try {
 			BufferedReader bufferedBackpackReader = new BufferedReader(new FileReader(backpackFile));
-			Backpack backpack = backpackGson.fromJson(bufferedBackpackReader, Backpack.class);
-			return backpack;
+			return backpackGson.fromJson(bufferedBackpackReader, Backpack.class);
 		} catch (FileNotFoundException e) {
 			Log.d(TAG, "Could not find backpack file!");
 			return new Backpack();
 		} catch (JsonSyntaxException | JsonIOException jsonException) {
 			Log.d(TAG, "Could not load backpack file! File will be deleted!");
-			backpackFile.delete();
+			deleteBackpackFile();
 			return new Backpack();
 		}
+	}
+
+	public boolean deleteBackpackFile() {
+		File backpackFile = new File(buildPath(DEFAULT_ROOT, BACKPACK_DIRECTORY, BACKPACK_FILENAME));
+		if (!backpackFile.exists()) {
+			Log.d(TAG, "Backpack file does not exist!");
+			return false;
+		}
+		Log.d(TAG, "Deleting backpack.json");
+		return backpackFile.delete();
 	}
 
 	public boolean codeFileSanityCheck(String projectName) {
